feat: Add scripts to build and run Wine containers
The code changes include the addition of two shell scripts, `build-wine32.sh` and `build-wine64.sh`, which are used to build Docker containers for Wine 32-bit and 64-bit respectively. The scripts use the `docker build` command to create the containers with the specified tags and Dockerfiles.

Additionally, two new files, `wine32` and `wine64`, are added. These files define the base images for the Wine containers, using Ubuntu 22.04 as the base and installing Wine 32-bit and 64-bit respectively.

This commit introduces the necessary infrastructure to build and run Wine containers, enabling users to easily set up and use Wine for compatibility testing and running Windows applications on Linux.